Move everything from SqlUtils into DbUtils
